Fix table rendering
Import the minimal amount of code from Bootstrap in order to render tables
nicely in posts, which is necessary by the latest post on sandboxfs.

This is a bit of a hack: we should instead be building Bootstrap from its
SASS sources and then using a small custom rule to inherit those rules into
the elements that need them... but this looks non-trivial.