Make the blog's front page readable
Our blog posts tend to be long... very long.  The previous blog index
page was showing the full posts which made it completely unreadable.
It was impossible to quickly skim through the page looking for what
the newest posts are, as the post titles would be "clobbered" by all
the content.

This change makes the index page only show the post excerpts and adds
a little footer to each indicating how much longer they are and how
many comments they have.  This makes it trivial, at a quick glance,
to see what has recently been published.

This should also help the trends I have observed in which some people
hold onto posting a new entry because the new entry would eclipse
older content.