Display authors on the posts index and give them some more visibility
This change moves the information about post authorship to the post's
Front Matter instead of having it as a hardcoded paragraph at the end.
This allows programmatic access to the authorship information, which
we then use to display author names in the post index and more
prominently next to the post title.

To ensure author details are always consistent (name and URL), this
moves all people data to a table in _config.yml and references it
using usernames from the Front Matter.