Don't use the posts layout for non-posts
The posts layout should be able to assume that it's dealing with posts
so that it can unconditionally access things like the page's date.
The archive is not a post, so use the blog layout instead.

This partially fixes the breakage introduced by
8b286b2dca8dfdd1978b0de6d99ad7ac22faf221 as the changes in that CL did
assume that page.date was always defined.